# api/openapi.yaml
openapi: 3.0.3
info:
  title: Deep Cuts CRM API
  version: "1.0"
servers:
  - url: /
paths:
...
  /api/office/drivers:
    get:
      operationId: listDrivers
      tags: [routes]
      responses:
        "200": { description: drivers, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Driver" } } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes:
    post:
      operationId: createRoute
      tags: [routes]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RouteInput" } } } }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}:
    parameters:
      - { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getRoute
      tags: [routes]
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}/stops:
    parameters:
      - { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      operationId: addStop
      tags: [routes]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AddStopInput" } } } }
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}/stops/{stopId}:
    parameters:
      - { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } }
      - { name: stopId, in: path, required: true, schema: { type: integer, format: int64 } }
    delete:
      operationId: removeStop
      tags: [routes]
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}/sequence:
    put:
      operationId: reorderStops
      tags: [routes]
      parameters: [ { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/SequenceInput" } } } }
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}/out:
    post:
      operationId: routeOut
      tags: [routes]
      parameters: [ { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/routes/{routeId}/complete:
    post:
      operationId: routeComplete
      tags: [routes]
      parameters: [ { name: routeId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses:
        "200": { description: route, content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/office/stops/{stopId}/proof:
    get:
      operationId: getStopProof
      tags: [routes]
      parameters: [ { name: stopId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses:
        "200": { description: proof image, content: { image/png: { schema: { type: string, format: binary } }, image/jpeg: { schema: { type: string, format: binary } } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
...
    Driver:
      type: object
      required: [id, displayName]
      properties:
        id: { type: integer, format: int64 }
        displayName: { type: string }
...
    OrderStatus:
      type: string
      enum: [draft, confirmed, scheduled, delivered, finalized, cancelled]
...
    RouteInput:
      type: object
      required: [routeDate, driverUserId]
      properties:
        routeDate: { type: string }
        driverUserId: { type: integer, format: int64 }
        truckLabel: { type: string }
    Stop:
      type: object
      required: [id, routeId, orderId, sequence, status, hasProofImage, skipReason, driverNote, customerName, deliveryAddress, phone, contactName, deliveryNotes, orderStatus, needsReview, lineCount]
      properties:
        id: { type: integer, format: int64 }
        routeId: { type: integer, format: int64 }
        orderId: { type: integer, format: int64 }
        sequence: { type: integer, format: int64 }
        status: { type: string, enum: [pending, delivered, skipped] }
        deliveredAt: { type: string, format: date-time }
        proofType: { type: string, enum: [signature, photo, name] }
        proofName: { type: string, description: received-by name when proofType is name }
        hasProofImage: { type: boolean }
        skipReason: { type: string }
        driverNote: { type: string }
        customerName: { type: string }
        deliveryAddress: { type: string }
        phone: { type: string }
        contactName: { type: string }
        deliveryNotes: { type: string }
        orderStatus: { $ref: "#/components/schemas/OrderStatus" }
        needsReview: { type: boolean }
        lineCount: { type: integer, format: int64 }
    Route:
      type: object
      required: [id, routeDate, driverUserId, driverName, truckLabel, status, stops]
      properties:
        id: { type: integer, format: int64 }
        routeDate: { type: string }
        driverUserId: { type: integer, format: int64 }
        driverName: { type: string }
        truckLabel: { type: string }
        status: { type: string, enum: [planned, out, complete] }
        outAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        stops: { type: array, items: { $ref: "#/components/schemas/Stop" } }
    AddStopInput:
      type: object
      required: [orderId]
      properties:
        orderId: { type: integer, format: int64 }
    SequenceInput:
      type: object
      required: [stopIds]
      properties:
        stopIds: { type: array, items: { type: integer, format: int64 } }
...